requirements.txt

streamlit==1.23.0
pandas==2.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.6.0
psychopy==2023.1.3
python-dotenv==1.0.0
numpy
scipy
matplotlib